Change folder structure: We only have one message folder in which you find MessageSets, a MessageSet is a file which contains all (needed) commands and replies from 1 command set as in https://docs.oracle.com/javase/7/docs/platform/
This way the structure is more closely to the protocol definition and
thus easier to grasp.
